rotating body towards!

+++ src/main/java/dev/zwazel/internal/game/tank/TankConfig.java
@@ -2,6 +2,6 @@ package dev.zwazel.internal.game.tank;
 
 import dev.zwazel.internal.game.transform.Vec3;
 
-public record TankConfig(float moveSpeed, float bodyRotationSpeed, float turretRotationSpeed, float maxSlope,
+public record TankConfig(float moveSpeed, double bodyRotationSpeed, double turretRotationSpeed, float maxSlope,
                          Vec3 size) {
 }
