+++ src/main/java/dev/zwazel/internal/message/data/tank/RotateTankTurretCommand.java
@@ -5,5 +5,5 @@ import dev.zwazel.internal.message.MessageData;
 import lombok.Builder;
 
 @Builder
-public record RotateTankTurretCommand(Tank.RotationDirection direction, float angle) implements MessageData {
+public record RotateTankTurretCommand(Tank.RotationDirection direction, double angle) implements MessageData {
 }
